Added a packet size check to the OP_AAAction handle to prevent a zone crash.
Revert "Commented out AA handling to stop zone crashing until we get to working with them."

This reverts commit 7481eaa480ef58374e4ac55d61768a3e192cae00.

Fixed quest methods summoning full stacks by default.

Changed the name of the "charges" variable to "quantity" in SummonItem for clarity. We'll have to go through and change the name in the code as we come across them.

Removed some senseless comments in SummonItem.