SoulMark Window & related functions, plus a bugfix.
Implements:

/praise <charname> <accid> <reason>
/warn <charname> <accid> <reason>
/inquire <charname> <accid>

Inquire shows the window ONLY for <charname>
Warn gives a 'negative' message for player with name <charname>
Praise gives a 'positive' message for player with name <charname>

AccountName is for GM reference only. GMAccountName is never used, but kept for reference as it's part of the client structure.

Also, Fixed an issue where dynamic zones (and some statics under heavy load) could result in kicking the player offline in world, causing /who all to break.