Client Push Support for CombatDamage_Struct
Added client push components to the damage packets in
CombatDamage_Struct.  Magnitudes for push pulled from decompile of the
client.  Changed sequence to float, since it represents the float of
direction when associated with push.  Removed manual push from Mob::Stun
since it will be handled in the damage packets for the client and it
caused too much jumping backward of clients due to latency between
server and client.