Added lua methods to support multiquest. The bulk of the implementation is done in lua.
Updated #npcstats/#npcloot show to display if the item in the NPC's loottable is a quest item or not. Also cleaned up the appearance some.

Updated #npcloot add to allow a GM/admin to add an item to the loottable with a quest flag.

Added a new rule NPC:IgnoreQuestLoot. When true, NPCs will not drop quest items handed to them. Defaults to false.

Just a note to clarify: This commit does not break any existing quests. The use of this code in Lua is optional.