Implemented an assist aggro cap to prevent a single NPC from calling too many NPCs to assist at once. Added rule Combat:NPCAssistCap to change that cap.
A NPC's assist cap limit is on a timer, and will decrease every time the timer hits.

Added rule AlKabor:AllowTickSplit. When disabled it causes the assist cap limit to be cleared when a NPC loses aggro. (defaults to false, but should be set to true for era correct behavior.)

NPCs that are answering another NPC's call for help will no longer ask for help until they have been attacked.

Added a new NPC specialattack 43 that allows the NPC to always yell for help if able, ignoring the cap and combat rules.

Pets will now only answer calls for help from their master.

Added an "attacked" option to #listnpcs to display a list of NPCs currently in combat (have been attacked and have a hate list).

Changed the default NPC loot equip_slot to 22, which is general inventory.