Added logging type Boats. SQL is added automagically.
Added code to help place you on the boat while zoning to firiona and timorous. Having the boat name in your player profile sends you to incorrect coords in these zones. This is a workaround until it can be figured out why this happens. Interestingly enough, overthere does not have this issue and the client will find the boat and place you on it like in all other boat zones.

Moved a lot of the boat code to a single method, SendToBoat().

Uncommented a log that will hopefully gives us more information when the server recieves an opcode it was not expecting during the connecting phase.

Removed AA_NO_TARGET string as our client does not have it in its string file.