Finished the implementation of Disciplines I started in 0171eb93f261503cda350f73a1f8465b0e515231
Discs are now treated as spells by the server code, but not the client. The server will use the effects in the spell data for each discipline. The animation, damage, and action packets are not sent when a discipline is cast, so the client is not aware that a spell has even been cast. It will use the normal discipline packet for the client side effects.

This commit includes a required SQL that contains all of the discipline spell data. The data was pulled from the PEQ database, but most of the spells have been heavily altered so their functionality matches our era instead of how Live exists currently. Despite this, exact percentages and magic numbers from our time were really hard to come by, so in many cases they were guessed at. The data may need to be tweaked as it is play tested. With this new system, these tweaks and changes can be done in the database. There shouldn't be a need to change the server code.

A new effect formula (150) was added to support the resitant discipline.

The discbonuses code was removed, as it is no longer needed. Disciplines now use spellbonuses like normal spells do.

Finished adding the discipline cast strings, and altered the framework to support the spell system as well as fix some bugs found along the way.

The following changes came about while testing discs:

Added missing resist bonuses for Beastlord and Monk.

While adding the new formula, the existing spell formula handles were cleaned up. Ones our spells don't use were removed, and a few missing ones were added.

Changed the order of the resists and stats in #showstats, so they match the order they appear in the client. Removed the heroic stats from there as well.

If #showstats 1 is typed, a shorter version of stats will now display with only core stats listed. Using this on a NPC has the same output as #npcstats.

Important note: Only the server uses the spell data in this commit. A new spells_us.txt file does NOT need to be generated and distributed to the players. The client is not aware of any of this data and does not need to be.

Required SQL (already ran against TAK):
utils/sql/git/required/2015_08_26_disc_spells.sql