NPCs that do not have a quest, or don't have a trade event within their quest will now have their handins completely handled by the server code.
The NPC:ReturnNonQuestItems rule will determine if those NPCs will return items, or add them to their inventory. The default setting of true hands these items back to the player.

No drop items or any item that the NPC already has will be destroyed upon handing it in, if the handin is handled by the server code. If it is handled by Lua, then those items will be destroyed on the death of the NPC, before the item packets are created. Summoned pets are an exception to this, and can be given duplicate items. GM handins are completely exempt.

We are now marking items handed to a pet in their inventory.

Added #logs quiet which disables all gmsay logging until either the zone resets, or a log is manually enabled/all are reloaded.

Exported CanTalk() to Lua, which allows Lua scripts to check if the NPC is a race that should be able to speak a language.