Invisible spells will now be broken on misses and resisted spells.
Corrected the color and type of invisible messages when it begins to fade. Invis spells that do not last their full duration will no longer have a blinking icon.

Disarm will no longer skillup when the NPC doesn't have a weapon that can be disarmed. Corrected the success message to only occur when the weapon actually drops.

Feign Death will now fail if auto attack is on, and will be broken when a spell is resisted.