All skills will now use stats to determine how quickly (or slowly, if they have low stats) a skill will skillup.
I couldn't find a decent formula, so this is completely pulled from thin air. However, I did try to keep the reuslting chances to be in the realm of what they are now. Skills that are using the minimium variable (tracking, swimming) will not see any change from this commit.

Added a second rule, Character:MasterSkillUpModifier which works the same as Character:SkillUpModifier but only works when the player's skill value is 100 or higher.