Added more security options to salt.
This would have been great to have been brought to the attention of the person who developed it in the proper channels instead of a public forum not related to this project.
Security holes discussed by devs of any portion of this code outside the scope of private channels is pretty unethical.

That being said, This defaults to creating a salt.ini based off of your initial login.ini.
If you choose to use any other means, you can adjust precoded options by commenting and uncommenting defines in salt.cpp
As well as create your own. This should be pretty straight forward, as long as
std::string Saltme::Salt(std::string password)
returns something, it will use it.

Commenting all should return a usable "null" for those with dev machines.
If not, make the adjustment and commit it.
Cmake WILL need to be ran.