NPCs that are not a playable race will now return to their proper texture after an illusion wears off.
All entities will now return to their proper size after an illusion wears off.

Corrected the issue causing Halflings and Half Elves to sometimes be the wrong size.

Cleaned up the size code and condensed base heights to a single method, GetPlayerHeight()

Shrink will now work while an illusion is active, and it should be less buggy in general.

Fixed the Disintegrate message sent when the spell is cast on a NPC that is too high in level.