Re-worked skillups to reflect formulas posted by Shendare here: http://www.eqemulator.org/forums/showpost.php?p=243531&postcount=20
This formula matches up with other sources, and is close to what we are currently using for tradeskills which came from eqtraders.

All skills now have a difficulty rating between 1 and 15 (15 being the hardest) that determines how difficult they are to skillup.

The above mentioned difficulty values are now stored in the skills_difficulty database table. This has the benefit of not only being able to change these skills on the fly, but we can now also set skills individually that in the old code were clumped together. (Weapon skills, Monk skills, spell skills, numchuku skills, bow hunting skills, computer hacking skills)

Skills that can succeed or fail and can skillup on failure will now take their success into account for the calculation.

Added command #skills. This can be used to view a player's current skill, their cap, and all the skills difficulties. It can also reload or change skill difficulties.

Important note: The old values EQEmu used were pulled from thin air. So, we are basically starting over. Feedback will be necessary to get all the skills leveling at the proper rate.

It is also important to note that tradeskill and language skillups are NOT affected by this change.

Corrected the name of the TradeskillSkillUpModifier rule. (Not sure how this was even compiling.) Moved LangSkillUpModifier and SkillUpModifier to the Skills category so all are uniform.

------------

Fixed the recourse on Greenmist and Soul Well.

Required SQL: (Already run against TAK)
utils/sql/git/required/2015_10_01_skill_difficulty.sql