Improved how weight is calculated server side. Implemented coin purses (server side - the client end was already working.)
Weight will now be recalculated when coin is moved to and from the bank.

Added IsEncumbered() to tell us when the server thinks the client is emcumbered.

The bestz adjustment at zone-in will be skipped when a player is encumbered.

Added some pet info and IsEncumbered to #showstats.

Fixed a couple of minor issues with code in previous commits of mine.