legacy table backup query added.
IF your queryserv crashes it is due to previous code not entering values in one or more of the tables that match it's companion table.
This will dump what you currently have to a textfile onto the system that the server is on with the name of the table and a .txt extension.

Redated the QS rule update sql to match the build date submitted to dev.