Spinstun duration revert
The previous spinstun duration change was put in because the spell data
had 1s in the effect value field, which I took to mean duration in ticks,
and the code wanted miliseconds.  However after examining logs, it's clear
spinstuns (at least on PCs) were lasting 7500 miliseconds.  Checking spell
data histories on allakhazam indicate that Sony changed the spell data to
7500 at one point.  AK seems to have also been 7500.

You will need to edit your databases and put in 7500 for the effect value
for spinstuns to work correctly.